pdf-drawing-diff/pyproject.toml
[project]
name = "pdf-drawing-diff"
version = "0.1.0"
description = "Skill source for multi-PDF drawing difference analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26.0",
    "opencv-python-headless>=4.8.0",
    "pillow>=10.0.0",
    "pymupdf>=1.24.0",
    "python-dotenv>=1.2.2",
    "requests>=2.28.0",
    "rich>=15.0.0",
    "zai-sdk>=0.2.2",
]
